Move router types to /types to keep consistency
The rest of the packages now export types in the types.ts file of the
root.
