+++ .changeset/mighty-books-doubt.md
@@ -1,0 +1,7 @@
+---
+"@frontity/head-tags": patch
+"@frontity/router": patch
+"@frontity/tiny-router": patch
+---
+
+Move router types to @frontity/router/types file to match the placement of the rest of the packages.
