+++ packages/head-tags/package.json
@@ -24,12 +24,12 @@
     "url": "https://github.com/frontity/frontity/issues"
   },
   "dependencies": {
-    "frontity": "^1.5.0"
+    "frontity": "^1.5.0",
+    "@frontity/router": "^1.0.17",
+    "@frontity/source": "^1.1.0"
   },
   "devDependencies": {
     "@frontity/connect": "^1.0.4",
-    "@frontity/router": "^1.0.17",
-    "@frontity/source": "^1.1.0",
     "@frontity/wp-source": "^1.5.0"
   }
 }
